# api/swagger/swagger.yaml
basePath: /
host: localhost:8888
info:
  contact:
    email: kartashov_egor96@mail.ru
    name: mail
    url: /
  description: eShopCDN
  license:
    name: MIT
    url: http://www.apache.org/licenses/MIT.html
  termsOfService: /
  title: eShopCDN Swagger Doc
  version: "1.0"
paths:
  /api/v1/cdn/download/{filename}:
    get:
      consumes:
      - application/json
      description: download file by name
      parameters:
      - description: Filenam
        in: path
        name: filename
        required: true
        type: string
      produces:
      - application/json
      responses: {}
      summary: "Allow \t\t\t\t\t\tdownload file"
      tags:
      - Files Controller
  /api/v1/cdn/file/{id}:
    delete:
      consumes:
      - application/json
      description: delete file by name
      parameters:
      - description: "123"
        in: header
        name: x-api-key
        required: true
        type: string
      - description: Filename
        in: path
        name: filename
        required: true
        type: string
      produces:
      - application/json
      responses: {}
      summary: Allow delete file
      tags:
      - Files Controller
  /api/v1/cdn/file/{id}/replace:
    post:
      consumes:
      - multipart/form-data
      description: replace file
      parameters:
      - description: "123"
        in: header
        name: x-api-key
        required: true
        type: string
      - description: File
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses: {}
      summary: Allow to replace file
      tags:
      - Upload Controller
  /api/v1/cdn/files:
    post:
      consumes:
      - application/json
      description: show paginated files
      parameters:
      - description: "123"
        in: header
        name: x-api-key
        required: true
        type: string
      - description: Page
        in: query
        name: page
        type: integer
      - description: Page
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses: {}
      summary: Allow get paginated files
      tags:
      - Files Controller
  /api/v1/cdn/metadata/{filename}:
    get:
      consumes:
      - application/json
      description: show file metadata
      parameters:
      - description: "123"
        in: header
        name: x-api-key
        required: true
        type: string
      - description: Filename
        in: path
        name: page
        required: true
        type: string
      produces:
      - application/json
      responses: {}
      summary: Allow get file metadata
      tags:
      - Files Controller
  /api/v1/cdn/upload:
    post:
      consumes:
      - multipart/form-data
      description: upload files
      parameters:
      - description: "123"
        in: header
        name: x-api-key
        required: true
        type: string
      - description: Files
        in: formData
        name: files
        required: true
        type: file
      produces:
      - application/json
      responses: {}
      summary: Allow upload multiple files
      tags:
      - Upload Controller
  /api/v1/livez:
    get:
      consumes:
      - application/json
      description: Health check service
      produces:
      - application/json
      responses:
        "200":
          description: OK
      summary: Health check service
      tags:
      - Health Controller
  /api/v1/readyz:
    get:
      consumes:
      - application/json
      description: Ready check service
      produces:
      - application/json
      responses:
        "200":
          description: OK
      summary: Ready check service
      tags:
      - Health Controller
schemes:
- http
swagger: "2.0"